General preferences form fixes
Only save mod_time for general prefs;
General prefs form requires the 2.x version of the jump module for the preference set dropdown menu. However, specific versions cannot be set in module .info dependencies. When drush installs dependencies, it installs the wrong version if left to its own devices. We must include the 2.x version of jump in the repo as an override.