VBOX: Disable USB support in virtual machines by default.
On Linux the currently signed-in user has to be added to the vboxusers group in order to handle bridging USB devices.  If the user has not taken that step virtual box throws a warning that looks like an error.  This caused vboxwrapper to believe that the VM had failed to start.  Since we do not know of any VM enabled problem that needs to communicate with an external USB device, just disable the feature entirely.