BOINC environment configuration
Fixed redirect loop problem when boinc_standard has not yet been reverted and is stuck on the garland theme. Set to boinc theme by default.
Removed configuration of site_name and site_mail, which are set during install. Also removed site_mission and site_slogan, which are not used in the theme and should not be controlled by the feature.