Login redirect on access denied
The r4032login module needed a patch to work for sites served in subdirectories, so this has been added to boinc/modules/contrib as a patched third party module.
Updated the standard page template so the login link includes a destination parameter to return the user to the current page after login.