Private message redirect issue
The privatemsg module contains a "feature" that prevents an anonymous user from knowing if a message thread exists or not by always returning a 404 error. This often breaks expected functionality when using links in email notifications to reach a new message. Replace this module with a patched module that returns 403 when appropriate.