Task details incorrect
Functions for converting status IDs to messages require objects as arguments, so variables were set to the proper format.
Dates were showing "e" instead of the timezone. This is a matter of the custom date format in Drupal not respecting the proper PHP "e" symbol for timezone. Changed to use "T" instead.
The task filter for Invalid tasks was broken in the view. The Views Or module is installed, so the reason for this is unknown. The filter was recreated to solve this.