VBOX: Enable the use of hardware acceleration on single core machines.
This requires BOINC 7.2.18 or better and for BOINC to have not have detected VM extensions being disabled in the BIOS.

T4T has stopped handing out multi-core work.  Disabling the use of hardware extensions on a machine that supports them might be leading to performance problems for other VMs running on the same machine.