User recent activity
Added a compound view that displays a summary of all of a user's forum posts and other content. This requires some fancy footwork with hook_views_pre_execute() to allow the view to include both node content and comments. A link to this view was added to the profile with the user's post count.

(DBOINCP-77)