client: Give boinc_master full control of any process spawned using the service token created by the client for boinc_project.
This will allow the core client to kill VirtualBox VM's launched indirectly by vboxwrapper.  Vboxwrapper launches vboxsvc.exe which launches vboxheadless.exe.  This should also take care of the core client being able to kill child processes of the regular wrapper as well.  I don't know the full scope of this type of issue?  Maybe the default ACLs for a process changed within the last couple of versions of Windows.