Client: fix bug that caused lots of spurious "no shared memory segment" msgs
ACTIVE_TASK_SET::suspend_all() originally skipped tasks
in states other than PROCESS_EXECUTING.
I took this out in commit 47b4d6b because - for example -
a GPU task might be suspended due to CPU throttling,
and therefore left in memory,
but if it's then suspended for some other reason,
it must be removed from memory.

However, this change was overkill - it causes tasks for which
no process exists to be suspended, resulting in the spurious msgs.

Solution: skip tasks in states other than PROCESS_EXECUTING
and PROCESS_SUSPENDED.