client: Default to /usr/bin when looking for VboxManage.
Depending on the Linux distro it can be in many different locations, even if it is installed in a different location there is normally a symbolic link in /usr/bin which points to the real deal.  Vboxwrapper should be able to run it successfully because it'll be in the standard search path.