Fix forum privileges
The Forum Access module was incorrectly defining the rid field as INT(11) rather than INT(10) unsigned. Because INT(11) isn't actually any bigger than INT(10), dropping the unsigned declaration causes half of the available rid space to be lost.

Until the module is fixed by the maintainer, it must be overridden locally. A database update has been added to the install file to redeclare the rid field as INT(10) unsigned.

(DBOINCP-91)