Versions for patched contrib modules
Further added a "-dev" flag to the revision number to all contrib modules that are patched locally. This is required to prevent Drupal from assuming that the recommended release with the same major and minor version numbers should be installed. The "-dev" flag forces the update process to look at the datestamp of the module as well as the version information and uses that to determine whether a recommended release is actually newer.