Comment quote preview layout bug
When quoting another comment, an inconsistency in how form buttons are defined causes the buttons to render below the quoted post rather than below the form. This has been updated to be consistent, but further caused another layout bug that would somehow move the form above the preview of the post. This has been worked around also.

(DBOINCP-93)