Client: don't abort jobs if WSS > workunit.rsc_memory_bound
On second thought, I don't think this is change was desirable.
Currently many projects underestimate rsc_memory_bound,
so lots of jobs would be aborted.
I don't want to force projects to give accurate estimates of RAM usage.
That may not even be possible.
And we already have mechanisms for aborting jobs
for which WSS > available RAM,
and only scheduling sets of jobs whose total WSS is < available RAM