client: don't try to run OpenCL jobs on non-OpenCL GPUs
Suppose
- the host has 2 GPUs of same vendor; A is OpenCL capable, B isn't
- the volunteer sets "use_all_gpus" config flag
Then the client will try to run OpenCL jobs on B.
Depending on how the app is written,
it may run on B and fail, or run on A and overload A.

Solution: when assigning GPUs to OpenCL jobs,
check that the GPU instance is OpenCL capable.

Note: this problem would go away if we treated each GPU as a separate resource.