Remove remnants from XML when clearing preference sets
If a preference set is established for a user (e.g. Home, Work), it is stored in the database with a <venue> wrapper. If all of these sets are then cleared from the account, a <venue/> tag was remaining, which causes parser issues in BOINC. Remove the <venue> element when all preference sets are cleared.

(DBOINCP-108)