Client: improve task duration estimates for apps that don't report fraction done
The "static estimate" is wu.rsc_fpops_est/app_version.flops.
The problem is: what if the elapsed time exceeds this.
In this case we were returning elapsed time,
resulting in a "time remaining" of zero, which is bad.

Instead, use the same exponential model that we use to
estimate fraction done when it's not reported.
This has the advantages that:
- time remaining monotonically decreases
  (though potentially at a very slow rate)
- the combo of fraction done, elapsed time, and time remaining
  is consistent for apps that don't report fraction done