Set minimum size for main content on front page
The height of the main content pane on the front page is balanced with the stats and news panels in the right column. If the news column was too short, a scrollbar was being rendered on the main content so that it would be smaller than its own content. Set a minimum height for this content on the front page.

(DBOINC-133)