Fixed linking of profile nodes to user accounts
The uid field of profile nodes gets overwritten by the node_submit() function if the name property of the node is not set to the user's name. It may be the case that the call to node_submit() is not even necessary, but setting the node->name property allows it to work as is.

(DBOINCP-123)