scheduler: fix bug that broke enforcement of per-app job limits
The old scheduler worked as follows:
    scan jobs; for each job
        get_app_version
        do various checks
        add_job_to_reply
The check for per-app job limits was in get_app_version,
and the incrementing of per-app job count is in add_job_to_reply.

The new (score-based) scheduler works as follows:
    scan jobs; for each job
        get_app_version
        add to list
    sort list by score
    scan list; for each job
        do various checks
        add_job_to_reply

So the limit check (in get_app_version) was ineffective
because it happens before we've incremented counts.

Fix: do the limit check (also) in the "scan list" loop.

Bigger picture: we need to restructure app version selection;
job limit enforcement doesn't belong there any more.