client: add notion of sticky file lifetime
If a <file_info> in scheduler RPC reply specifies <sticky_lifetime>,
the client will calculate a "sticky_expire_time",
store it in the client state file,
and make the file unsticky when that time is reached.

Note: if a later RPC reply includes the same file,
the client will update sticky_expire_time.
So if a file is used repeatedly it won't get expired.