admin web: in "result summary in last X", don't include delayed-file-delete jobs
The config option <delete_delay_hours> lets you defer file deletion.
If you set this, say, for 10 days,
then the "results summary in last 24 hours" page will show
jobs that finished 10 days ago and were file-deleted in the last day.
This is generally not what you want.
Change things so that in this case file-deleted results
are treated as if their mod time were 10 days earlier.

Note: this means that the counts of file-delete states
won't be meaningful in this case.