client: increase limit on coproc instances from 31 to 64
We were using an int bitmap to store flags for the instances of a coproc.
Furthermore, because of the use of 2^n-1 to generate a bitmap of 1s,
the limit on instances was 31.

Use a long long for the bitmap instead, and don't use 2^n-1.
This increases the limit to 64.