Fixed delete team function
The remove team link was not working, as it required that all team members be removed first. Fixed so that all but one member must be removed, as that one member must remain in order for the team to be removed! Also, remove the user from the team when removing the team.

(DBOINCP-183)