client: preserve min RPC time across client restart
My 18 Oct 2013 commit caused the client to clear PROJECT::min_rpc_time
on startup.
This causes project-requested RPC delays to be ignored if
the client is restarted.
This is probably undesirable, so I unded that change.

Note: min_rpc_time is used for both
a) project-requested RPC delay
b) exponential backoff due to server down, not work, etc.

To me it makes sense to clear b) on restart, but not a).
Maybe at some point we should separate these.