web: don't show non-English translation incorrectly
The translation code (translation.inc) had a special case
where if the first accepted language was "en", it would show English.
Problem: if the accepted languages are, say, "en-us", "en", "es"
then you'll see Spanish.
Fix: check for both "en" and "en_*" in the special case.

The logic of this code is still flawed, however.
It builds a list $languages_in_use of accepted languages
for which a translation file exists,
and to translate a word it tries these in sequence.
So a page could have translations from several languages.

What it should do is scan the accepted languages until it
finds one that either is English or for which a translation file exists.
Then use that one.
(there could be a check that skips languages for which
the translation is very incomplete).