VBOX: If polling for the current VM state fails for any reason, like vboxsvc crashing, do a temp exit for 24 hours.
Before COM, polling for the current state wasn't as critical.  If the poll operation failed, we would try again in one second anyway.

After COM, it is used as a test to see if the communication channel between the wrapper and VirtualBox is still operational.  If not, temp exit and hope things work the next attempt.