web: changes to support MySQL strict mode
In "strict mode", inserts fail if they don't supply values
for columns with no default defined in the schema
(in non-strict mode, 0 and '' are implicit defaults).
Starting with MySQL version 5.6, strict mode is the default.
This breaks some of the BOINC web code,
which does inserts without giving values to some columns.

There are two ways to solve this:

1) change the schema to give defaults everywhere
2) change the PHP code to supply values for more columns.

I'm using 1) in some cases and 2) in others.
This commit fixes some of the errors; there are others.