client: fix bug that caused delay in job cleanup
If a job has an output file with <copy_file> and <optional>,
and it doesn't create the file,
then the call to boinc_rename() (to move it to the project dir) fails,
and we back off and retry.

Solution: in boinc_rename(), if the rename fails,
check if the file exists, and if it doesn't then don't retry.

Also:
- when writing client messages, use the actual current time
  (dtime()) rather than client_state.now.
- write log msgs when output file renames fail