web: make "email friends" feature work without project configuration
Previously the feature that lets people send email about the project
to their friends only worked if the project supplied a message.
Remove this restriction; supply a default message.
Also make it work with or without PHPMailer.

Projects can still supply a default subject and message;
the user can now edit these.

Remove the ability to send HTML email
and to macro-substitute names in the message.