client: fix problem where app processes inherit open files from async copies
Github issue #1388 describes a situation where a wrapper can't exec
an executable file because there's an open file descriptor to it.
This happens only if the file is > 10 MB so that async copy is used.

The reason for this (which Rom identified) is as follows.
Suppose several such tasks start around the same time.
The first async copy finishes and the wrapper is started by fork/exec.
The process inherits open fds for all the other copies.
When they finish and are started, they won't be able to exec.

The fix is to use boinc_fopen() instead of fopen() in ASYNC_COPY.
The former sets its fd to close-on-exec.

In general the client should always use boinc_fopen() instead of fopen();
in addition to close-on-exec, it also does retry.
I changed a couple of other fopen()s also.