VBOX: Add code to handle search path modification for Linux and Mac.
Something has changed on OS X that prevents /usr/local/bin from being in the default search path.  It is when launched from a terminal window, but not when BOINC launches vboxwrapper.  I think recent versions of VirtualBox changed where it stores VboxManage on install.  bbad0d9d recently addressed this in the CC.