web: don't accidentally write to read-only DB replica
BOINC lets you maintain a read-only DB replica,
and certain read-only web pages (like top_users.php)
use the replica in preferences to the master.
But it turns out that these pages aren't actually read-only
because of the too-clever forum_preferences logic.
So sometimes it would add forum_preferences records to the replica,
leading to errors it later copies the same item from the master.

Solution: skip the forum preferences logic if we're using the replica.

We could make this more bulletproof by putting in checks
in insert(), update(), delete() etc.