client: increase app-exit timeout from 15 to 60
This is the interval between when the client sends a "quit" message
to when it kills the app via TerminateProcess or SIGKILL.

Apparently VM apps that do their own state-saving
(i.e. that don't use snapshot-based checkpointing)
can take more than 15 seconds to do so.
Hopefully 60 is enough.

Rom suggested making this interval shorter in the case where
the OS is shutting down.
I don't think this is necessary since the OS kill everything anyway
after some period (5 sec in the case of Win 10).