Server: fix sched/start to work with older python version
An issue with unicode strings in python 2.4 and 2.6 (and possibly 2.5) prevents shlex to split the command which leads to the daemon or task not starting. The unicode issue seems to be fixed in python 2.7. The exact error message is: "TypeError: execv() argument 1 must be (encoded string without NULL bytes), not str".

See: https://github.com/vinodc/gitlab-webhook-branch-deployer/issues/1