wrapper: reclassify the wrapper as a console application on Windows.
This was a hold over from the Win9x days when a console application window could not be hidden by the processed that launched it.  Since we no longer need to support Win9x we can get rid of this hack.

NOTE: At some point parameter parsing and processing the job file was moved ahead of the diagnostics initialization code.  This had the unfortunate consequence of making errors that happened before the diagnostics code initialized disappear.

Switching wrapper to a console application will mean that launching it from a console window will cause the console window to block and display anything written to stderr/stdout regardless of whether the diagnostics code is initialized.