client/Manager: show correct app version in multi-platform case
The GUI RPC element for a result didn't include the platform name.
So in cases where there were app versions with the same
version # and plan class (but different platforms)
the Manager would just show the first one
(e.g. the "executable file" item in Task Properties),
which was not necessarily correct.